Finished Logic and cleanup

+++ GameLogic/src/Game/Player/Player.java
@@ -5,9 +5,6 @@ import Game.*;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.ListIterator;
-
-import static Game.PlayerTurn.*;
 
 /**
  * Created by ido on 18/08/2016.
@@ -121,7 +118,7 @@ public class Player {
             }
         }
         playerWon = (numberOfCorrcetSquares == numberOfBoardSquares);
-        statistics.setScore(numberOfCorrcetSquares / numberOfBoardSquares);
+        statistics.setScore((double) numberOfCorrcetSquares / (double) numberOfBoardSquares);
     }
 
     public void printMoveHistory()
